(Issue #43) Clean-up code as part of code review.

+++ src/Milhoja_GridAmrex.cpp
@@ -48,8 +48,8 @@ bool GridAmrex::domainDestroyed_   = false;
 GridAmrex::GridAmrex(void)
     : Grid(),
       AmrCore(),
-      unk_{static_cast<unsigned int>(max_level+1)},
-      fluxes_{static_cast<unsigned int>(max_level+1)},
+      unk_{max_level+1},
+      fluxes_{max_level+1},
       bcs_{1},
       comm_{GridConfiguration::instance().mpiComm},
       nBlocksX_{GridConfiguration::instance().nBlocksX},
